fix(sceneview-core): review follow-ups — Long SPZ offsets, clamp inflate pre-allocation, honest v3 claim (#2684 review)
W1: SPZ structure-of-arrays offsets now use Long arithmetic so a lying header
(numPoints ~2^31/9) trips the up-front truncation check instead of wrapping
Int past it (mirrors PlyParser's Long 'needed' guard).
W2: the v3 smallest-three KDoc claim is softened — round-trip-validated,
anticipated layout pending a real-world exporter fixture (v2 established).
W3: gzip pre-allocation is clamped to input.size * 1024 so a ~30-byte stream
claiming ISIZE=256MB can't force that allocation up-front (memory-amplification
DoS); ensureCapacity grows organically past the clamp, ISIZE still verified.
NITs: SplatCloud KDoc states array ownership (same sentence mirrored on the
P1b branch to keep the file byte-identical); dead logOf() test helper removed.

Gates re-run: androidTest+jsTest 0 failed, detekt clean.

+++ sceneview-core/src/commonMain/kotlin/io/github/sceneview/core/splat/SpzParser.kt
@@ -6,7 +6,9 @@ import kotlin.math.sqrt
 
 /**
  * Parser for Niantic **SPZ** files, the gzip-compressed 3D Gaussian Splatting interchange format
- * (https://github.com/nianticlabs/spz). Decode formulas mirror the reference `load-spz.cc` exactly.
+ * (https://github.com/nianticlabs/spz). Decode formulas follow the reference `load-spz.cc`; the v3 smallest-three
+ * bit layout is validated by encode/decode round-trip only — pending a real-world
+ * exporter fixture, treat v3 as an anticipated layout (v2 is the established format).
  *
  * Supported: the legacy gzip container, version **2** (`first-three` quaternion, 3 bytes) and
  * version **3** (`smallest-three` quaternion, 4 bytes) — the formats exported by Scaniverse, Marble,
@@ -59,14 +61,16 @@ internal object SpzParser {
         val shDim = dimForDegree(shDegree)
         val rotStride = if (smallestThree) 4 else 3
 
-        // Structure-of-arrays offsets, in serialization order.
-        var offset = HEADER_SIZE
-        val positionsOffset = offset; offset += numPoints * 9
-        val alphasOffset = offset; offset += numPoints
-        val colorsOffset = offset; offset += numPoints * 3
-        val scalesOffset = offset; offset += numPoints * 3
-        val rotationsOffset = offset; offset += numPoints * rotStride
-        offset += numPoints * shDim * 3 // sh (ignored, but must be present)
+        // Structure-of-arrays offsets, in serialization order. Long arithmetic:
+        // a lying header (numPoints ~2^31/9) must trip THIS check, not wrap Int
+        // and sneak past it (same defense PLY's `needed` uses).
+        var offset = HEADER_SIZE.toLong()
+        val positionsOffset = offset.toInt(); offset += numPoints.toLong() * 9
+        val alphasOffset = offset.toInt(); offset += numPoints.toLong()
+        val colorsOffset = offset.toInt(); offset += numPoints.toLong() * 3
+        val scalesOffset = offset.toInt(); offset += numPoints.toLong() * 3
+        val rotationsOffset = offset.toInt(); offset += numPoints.toLong() * rotStride
+        offset += numPoints.toLong() * shDim * 3 // sh (ignored, but must be present)
         if (offset > d.size) {
             splatError("SPZ: data truncated (need $offset bytes, have ${d.size})")
         }
